feat: double to uint16_t

+++ GP2Y0E03-Arduino.cpp
@@ -14,7 +14,7 @@ GP2Y0E03::GP2Y0E03(uint8_t pin)
     GP2Y0E03::setMinDist(MIN_DIST);
 }
 
-double GP2Y0E03::read()
+uint16_t GP2Y0E03::read()
 {
     uint16_t value = 0;                                     // Valor analógico bruto
     for (uint8_t i = 0; i < GP2Y0E03::getNumReading(); i++) // Somatório dos valores analógicos
@@ -25,7 +25,7 @@ double GP2Y0E03::read()
 
 uint8_t GP2Y0E03::detected()
 {
-    double read = GP2Y0E03::read();
+    uint16_t read = GP2Y0E03::read();
     _detected = (read <= GP2Y0E03::getMaxDist() && read >= GP2Y0E03::getMinDist());
     return _detected;
 }
@@ -35,12 +35,12 @@ uint8_t GP2Y0E03::getNumReading()
     return _numReading;
 }
 
-double GP2Y0E03::getMaxDist()
+uint16_t GP2Y0E03::getMaxDist()
 {
     return _maxDist;
 }
 
-double GP2Y0E03::getMinDist()
+uint16_t GP2Y0E03::getMinDist()
 {
     return _minDist;
 }
@@ -50,7 +50,7 @@ void GP2Y0E03::setNumReading(uint8_t numReading)
     _numReading = numReading;
 }
 
-void GP2Y0E03::setMaxDist(double maxDist)
+void GP2Y0E03::setMaxDist(uint16_t maxDist)
 {
     if (maxDist >= MIN_DIST && maxDist <= MAX_DIST)
         _maxDist = maxDist;
@@ -58,7 +58,7 @@ void GP2Y0E03::setMaxDist(double maxDist)
         _maxDist = MAX_DIST;
 }
 
-void GP2Y0E03::setMinDist(double minDist)
+void GP2Y0E03::setMinDist(uint16_t minDist)
 {
     if (minDist >= MIN_DIST && minDist <= MAX_DIST)
         _minDist = minDist;
